search bars styliing and integration
